Revert "Add LLDB debugger support to swift-test command (#9085)"
This reverts commit e1a62e1fa07bb19ed9149626b5c40fef8f377bbe.
