--- Fixtures/Miscellaneous/TestDebuggingMultiProduct/Sources/LibA/LibA.swift
@@ -1,7 +1,0 @@
-public struct LibA {
-    public init() {}
-
-    public func greet() -> String {
-        return "Hello from LibA"
-    }
-}
