--- Fixtures/Miscellaneous/TestDebuggingMultiProduct/Sources/LibB/LibB.swift
@@ -1,7 +1,0 @@
-public struct LibB {
-    public init() {}
-
-    public func greet() -> String {
-        return "Hello from LibB"
-    }
-}
