+++ Sources/Commands/SwiftRunCommand.swift
@@ -19,6 +19,7 @@ import PackageModel
 import SPMBuildCore
 
 import enum TSCBasic.ProcessEnv
+import func TSCBasic.exec
 
 import enum TSCUtility.Diagnostics
 
@@ -152,8 +153,7 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
                 fileSystem: swiftCommandState.fileSystem,
                 executablePath: interpreterPath,
                 originalWorkingDirectory: swiftCommandState.originalWorkingDirectory,
-                arguments: arguments,
-                observabilityScope: swiftCommandState.observabilityScope
+                arguments: arguments
             )
 
         case .debugger:
@@ -183,17 +183,12 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
                         fileSystem: swiftCommandState.fileSystem,
                         executablePath: debuggerPath,
                         originalWorkingDirectory: swiftCommandState.originalWorkingDirectory,
-                        arguments: debugger.extraCLIOptions + [productAbsolutePath.pathString] + options.arguments,
-                        observabilityScope: swiftCommandState.observabilityScope
+                        arguments: debugger.extraCLIOptions + [productAbsolutePath.pathString] + options.arguments
                     )
                 } else {
                     let pathRelativeToWorkingDirectory = productAbsolutePath.relative(to: swiftCommandState.originalWorkingDirectory)
                     let lldbPath = try swiftCommandState.getTargetToolchain().getLLDB()
-                    try safeExec(
-                        path: lldbPath.pathString,
-                        args: ["--", pathRelativeToWorkingDirectory.pathString] + options.arguments,
-                        observabilityScope: swiftCommandState.observabilityScope
-                    )
+                    try exec(path: lldbPath.pathString, args: ["--", pathRelativeToWorkingDirectory.pathString] + options.arguments)
                 }
             } catch let error as RunError {
                 swiftCommandState.observabilityScope.emit(error)
@@ -212,8 +207,7 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
                     fileSystem: swiftCommandState.fileSystem,
                     executablePath: swiftInterpreterPath,
                     originalWorkingDirectory: swiftCommandState.originalWorkingDirectory,
-                    arguments: arguments,
-                    observabilityScope: swiftCommandState.observabilityScope
+                    arguments: arguments
                 )
                 return
             }
@@ -251,8 +245,7 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
                     fileSystem: swiftCommandState.fileSystem,
                     executablePath: runnerPath,
                     originalWorkingDirectory: swiftCommandState.originalWorkingDirectory,
-                    arguments: arguments,
-                    observabilityScope: swiftCommandState.observabilityScope
+                    arguments: arguments
                 )
             } catch Diagnostics.fatalError {
                 throw ExitCode.failure
@@ -301,8 +294,7 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
         fileSystem: FileSystem,
         executablePath: AbsolutePath,
         originalWorkingDirectory: AbsolutePath,
-        arguments: [String],
-        observabilityScope: ObservabilityScope
+        arguments: [String]
     ) throws {
         // Make sure we are running from the original working directory.
         let cwd: AbsolutePath? = fileSystem.currentWorkingDirectory
@@ -311,7 +303,7 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
         }
 
         let pathRelativeToWorkingDirectory = executablePath.relative(to: originalWorkingDirectory)
-        try safeExec(path: executablePath.pathString, args: [pathRelativeToWorkingDirectory.pathString] + arguments, observabilityScope: observabilityScope)
+        try execute(path: executablePath.pathString, args: [pathRelativeToWorkingDirectory.pathString] + arguments)
     }
 
     /// Determines if a path points to a valid swift file.
@@ -332,6 +324,41 @@ public struct SwiftRunCommand: AsyncSwiftCommand {
             absolutePath = try AbsolutePath(cwd, validating: path)
         }
         return fileSystem.isFile(absolutePath)
+    }
+
+    /// A safe wrapper of TSCBasic.exec.
+    private func execute(path: String, args: [String]) throws -> Never {
+        #if !os(Windows)
+        // Dispatch will disable almost all asynchronous signals on its worker threads, and this is called from `async`
+        // context. To correctly `exec` a freshly built binary, we will need to:
+        // 1. reset the signal masks
+        for i in 1..<NSIG {
+            signal(i, SIG_DFL)
+        }
+        var sig_set_all = sigset_t()
+        sigfillset(&sig_set_all)
+        sigprocmask(SIG_UNBLOCK, &sig_set_all, nil)
+
+        #if os(FreeBSD) || os(OpenBSD)
+        #if os(FreeBSD)
+        pthread_suspend_all_np()
+        #endif
+        closefrom(3)
+        #else
+        #if os(Android)
+        let number_fds = Int32(sysconf(_SC_OPEN_MAX))
+        #else
+        let number_fds = getdtablesize()
+        #endif /* os(Android) */
+
+        // 2. set to close all file descriptors on exec
+        for i in 3..<number_fds {
+            _ = fcntl(i, F_SETFD, FD_CLOEXEC)
+        }
+        #endif /* os(FreeBSD) || os(OpenBSD) */
+        #endif
+
+        try TSCBasic.exec(path: path, args: args)
     }
 
     public init() {}
